Move PathEnumerator to models module
