+++ modules/pipeline/prefix-span-pattern-detector/src/main/kotlin/org/cafejojo/schaapi/pipeline/patterndetector/prefixspan/PatternDetector.kt
@@ -2,6 +2,7 @@ package org.cafejojo.schaapi.pipeline.patterndetector.prefixspan
 
 import org.cafejojo.schaapi.models.GeneralizedNodeComparator
 import org.cafejojo.schaapi.models.Node
+import org.cafejojo.schaapi.models.PathEnumerator
 import org.cafejojo.schaapi.pipeline.Pattern
 import org.cafejojo.schaapi.pipeline.PatternDetector
